# api/openapi.yml
openapi: 3.0.3
info:
  title: Frontier API
  description: |
    Comprehensive API ecosystem for Frontier AI capabilities including visual design, 
    self-improvement, code quality analysis, image generation, and audio/video processing.
    
    ## Authentication
    This API uses JWT tokens for authentication. Obtain a token from the `/auth/login` endpoint
    and include it in the `Authorization` header as `Bearer <token>`.
    
    ## Rate Limiting
    Rate limits are enforced based on your subscription tier:
    - **Free**: 100 requests/hour
    - **Developer**: 1,000 requests/hour  
    - **Professional**: 10,000 requests/hour
    - **Enterprise**: Custom limits
    
    ## Error Handling
    All endpoints return standard HTTP status codes with detailed error messages in JSON format.
  version: 1.0.0
  contact:
    name: Frontier API Support
    url: https://api.frontier.ai/support
    email: support@frontier.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.frontier.ai/v1
    description: Production server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /visual-design/brand-identity:
    post:
      tags: [Visual Design]
      summary: Generate brand identity package
      operationId: generateBrandIdentity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrandIdentityRequest'
      responses:
        '200':
          description: Brand identity generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrandIdentityResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/RateLimit'
...
  /code-quality/scan:
    post:
      tags: [Code Quality]
      summary: Scan code for quality issues
      operationId: scanCodeQuality
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CodeQualityScanRequest'
      responses:
        '200':
          description: Code scan completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodeQualityScanResponse'
...
  /image-generation/assets:
    post:
      tags: [Image Generation]
      summary: Generate website assets from descriptions
      operationId: generateAssets
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssetGenerationRequest'
      responses:
        '200':
          description: Assets generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssetGenerationResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    BrandIdentityRequest:
      type: object
      required: [businessName, industry]
      properties:
        businessName:
          type: string
          example: TechCorp Solutions
        industry:
          type: string
          example: Technology
        description:
          type: string
          example: AI-powered business solutions
        targetAudience:
          type: string
          example: Enterprise businesses
        style:
          type: string
          enum: [modern, classic, minimalist, bold, elegant]
          example: modern
        colors:
          type: array
          items:
            type: string
          example: [blue, white, gray]
    BrandIdentityResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            logos:
              type: array
              items:
                type: object
                properties:
                  type:
                    type: string
                    example: primary
                  format:
                    type: string
                    example: svg
                  url:
                    type: string
                    example: https://api.frontier.ai/assets/logo-primary.svg
            colorPalette:
              type: object
              properties:
                primary:
                  type: string
                  example: "#2563eb"
                secondary:
                  type: string
                  example: "#64748b"
                accent:
                  type: string
                  example: "#f59e0b"
            typography:
              type: object
              properties:
                heading:
                  type: string
                  example: Inter
                body:
                  type: string
                  example: Open Sans

    # Additional schemas would continue here for all other endpoints...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid request parameters
            details:
              type: array
              items:
                type: string
              example: ["Email is required", "Password must be at least 8 characters"]
  responses:
    BadRequest:
      description: Bad request - invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RateLimit:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit per hour
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Requests remaining in current window
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Time when rate limit resets (Unix timestamp)
...